refactor to a py project
