+++ README.md
@@ -1,7 +1,16 @@
-# Large Language Modeling
+# LLM(Large Language Modeling)
+
+```python
+pip3 install --no-cache-dir -i https://mirrors.aliyun.com/pypi/simple/ --upgrade -e ".[data]"
+```
 
 # Pre-training
 
 # Fine-tuning
+1. Process data
+    ```python
+    python -m llm.data.process_chats --dataset PIPPA
+    ```
 
-# RL*F
+
+# Alignment
