+++ pyproject.toml
@@ -1,0 +1,33 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "llm"
+version = "0.1"
+description = "A platform for training and evaluating large language model based chatbots."
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Apache Software License",
+]
+dependencies = [
+    "setuptools_scm", "aiohttp", "fastapi", "httpx", "markdown2[all]", "nh3", "numpy",
+    "prompt_toolkit>=3.0.0", "pydantic<2,>=1", "requests", "rich>=10.0.0",
+    "shortuuid", "tiktoken", "uvicorn",
+]
+
+[project.optional-dependencies]
+data = ["lingua-language-detector"]
+
+
+[project.urls]
+"Homepage" = "https://github.com/congchan/llm"
+"Bug Tracker" = "https://github.com/congchan/llm/issues"
+
+[tool.setuptools.packages.find]
+exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]
+
+[tool.wheel]
+exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]
